refactor: split the charts structure
